chore: tidy sip lifecycle types
