--- api/assistant-api/sip/runtime/call_termination.go
@@ -1,14 +1,0 @@
-// Copyright (c) 2023-2025 RapidaAI
-// Author: Prashant Srivastav <prashant@rapida.ai>
-//
-// Licensed under GPL-2.0 with Rapida Additional Terms.
-// See LICENSE.md or contact sales@rapida.ai for commercial usage.
-
-package sip_runtime
-
-type CallTerminationResult string
-
-type CallTermination struct {
-	Result CallTerminationResult
-	Reason string
-}
